requirements.txt
flask==2.0.1
flask-cors==3.0.10
tensorflow==2.8.0
numpy==1.21.2
opencv-python==4.5.3.56
google-generativeai==0.3.0
python-dotenv==0.19.1
